Vertex_AI_Vector_Search/requirements.txt
langchain
langchain-community
langchain-core
langsmith
langgraph
langchain_google_genai
langchain_google_vertexai
langchain_google_community
